Einbau des Heatflux. Alpha Version. Tests laufen.

+++ src/MarDyn.cpp
@@ -199,7 +199,7 @@ int main(int argc, char** argv) {
 	}
 
 	/* processing command line arguments */
-	if ( options.get("legacy-cell-processor") ) {
+	if ( (int) options.get("legacy-cell-processor") ) {
 		simulation.useLegacyCellProcessor();
 		global_log->info() << "--legacy-cell-processor specified, using legacyCellProcessor" << endl;
 	}
